requirements.txt
asgiref==3.9.2
Django==5.2.6
sqlparse==0.5.3
python-decouple
pillow
djangorestframework
markdown 
django-filter 
drf-yasg
djangorestframework-simplejwt
django-mail-templated
djoser
black
flake8
pytest
pytest-django
Faker
celery
redis


    